+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## **Unreleased**
+## **Release v1.55.4 - 2026-08-10**
 **Hearthstone**:
 - The opponent's counters now account for Azalina Soulsever.\
   *Half of their deck is copied from yours, so counters for payoffs in your own deck now also show up for the opponent once they start counting.*